refactor: drop "pkg" directory from "internal"
Related to #44.
