+++ controlplane/internal/ffi/dp_config.go
