+++ controlplane/internal/gateway/logging_service.go
