+++ controlplane/internal/gateway/registry.go
