+++ controlplane/pkg/yncp/director.go
@@ -7,7 +7,7 @@ import (
 	"go.uber.org/zap"
 	"golang.org/x/sync/errgroup"
 
-	"github.com/yanet-platform/yanet2/controlplane/internal/pkg/gateway"
+	"github.com/yanet-platform/yanet2/controlplane/internal/gateway"
 	"github.com/yanet-platform/yanet2/controlplane/modules/route"
 )
